pyproject.toml
[project]
name = "geoquant"
version = "0.2.0"
description = "Geometric & Topological Embedding Analysis in MobileNetV3 under INT8 Quantization"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy",
    "scipy",
    "scikit-learn",
    "matplotlib",
    "pyyaml",
    "tqdm",
    "dvc",
    "dvc-gdrive",
    "mlflow",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]
torchao = [
    "torchao",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geoquant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true